Fixed titles in !daily and !keen
!daily used to say "Leaderboard for None:" on the top row.
Now it has a title with the year and day shown.
Altered the parameter name to be more in line with what it does, and commands pass in the full title now. (year -> title)
Fixed !keen having different title parity by moving the title inside the code block.

Fixed referencing undeclared variable "year".

"year" exists in [this PR](https://github.com/CSSUoB/advent-of-code-bot/pull/12)
This will need to be changed if/when both are merged to [the parent](CSSUoB/advent-of-code-bot)

Included !keen title in the code block

This fixes parity with the other commands, which have the leaderboard title at the top of their code blocks.